wip: refactor list command
